+++ functions/src/index.ts
@@ -1,1 +1,2 @@
-﻿export { getFirebaseAdminApp } from "./integrations/firebaseAdmin";
+export * from "./domain/roomMatchCommands";
+export { getFirebaseAdminApp } from "./integrations/firebaseAdmin";